Move infoButton to shared folder
